修 Decimal 比较二义: price/qty 合法性判断显式 cast double

+++ src/ts_case.cpp
@@ -235,7 +235,8 @@ inline void TsCase::onTrade(const MD::Trade &trade) {
   // 但按合法性过滤: 实盘 feed 存在 price=0 的非常规 print(原被 r1 filter 顺带
   // 挡掉; 2026-09-01 10:55 SNDK strat006 pnl 瞬间 +7606 = pos×init_px 即由此),
   // 会污染 OrderManager 标记价与 trade 类信号。
-  if (trade.price <= 0 || trade.quantity <= 0)
+  if (static_cast<double>(trade.price) <= 0.0 ||
+      static_cast<double>(trade.quantity) <= 0.0)
     return;
   msg_.type = enums::EventType::TRADE;
   msg_.price = static_cast<double>(trade.price);
